api/requirements.txt
azure-functions
azure-storage-blob
pandas
scikit-learn